split off persistence api
