+++ server/workers/persistence/requirements.txt
@@ -1,0 +1,14 @@
+flask
+flask-cors
+flask_sqlalchemy
+flask_restx
+Werkzeug
+marshmallow
+gunicorn
+redis
+hiredis
+aioredis
+pandas
+pyyaml
+flasgger
+psycopg2-binary
