+++ server/workers/persistence/src/config/example_settings.py
@@ -1,0 +1,6 @@
+BEHIND_PROXY = True
+SWAGGER_BASEPATH = ""
+DEFAULT_DATABASE = "dev"
+DATABASES = ["test"]
+ENV = "development"
+DEBUG = True
